+++ src/slic3r/GUI/GalleryDialog.cpp
@@ -519,7 +519,7 @@ void GalleryDialog::change_thumbnail()
 void GalleryDialog::select(wxListEvent& event)
 {
     int idx = event.GetIndex();
-    Item item { into_u8(m_list_ctrl->GetItemText(idx)), idx < m_sys_item_count };
+    Item item { into_u8(m_list_ctrl->GetItemText(idx)), size_t(idx) < m_sys_item_count };
 
     m_selected_items.push_back(item);
 }
